Warn that SMS not private
