+++ AndroidManifest.xml
@@ -91,7 +91,7 @@
         </activity-alias>
 
         <activity
-            android:name=".ui.permissioncheck.PermissionCheckActivity"
+            android:name=".ui.onboarding.OnboardingActivity"
             android:configChanges="orientation|screenSize|keyboardHidden"
             android:theme="@style/Theme.Compose" />
