+++ app/src/androidTest/kotlin/com/android/common/test/rules/AppTestRule.kt
@@ -1,6 +1,7 @@
 package com.android.common.test.rules
 
 import com.android.common.test.helpers.ShellCommandHelper
+import com.android.common.test.helpers.SmsWarningHelper
 import org.junit.rules.TestRule
 import org.junit.runner.Description
 import org.junit.runners.model.Statement
@@ -14,6 +15,7 @@ class AppTestRule : TestRule {
         return object : Statement() {
             override fun evaluate() {
                 val previousSmsRoleHolders = ShellCommandHelper.setupSmsDefaultRole()
+                val wasSmsWarningAcknowledged = SmsWarningHelper.acknowledgeSmsWarning()
                 var baseFailure: Throwable? = null
                 try {
                     base.evaluate()
@@ -21,6 +23,9 @@ class AppTestRule : TestRule {
                     baseFailure = throwable
                 } finally {
                     try {
+                        SmsWarningHelper.restoreSmsWarning(
+                            wasAcknowledged = wasSmsWarningAcknowledged,
+                        )
                         ShellCommandHelper.restoreSmsDefaultRole(
                             previousRoleHolders = previousSmsRoleHolders,
                         )
