+++ app/src/test/kotlin/com/android/messaging/ui/ActivityPermissionGateTest.kt
@@ -4,7 +4,7 @@ import com.android.messaging.ui.classzero.ClassZeroActivity
 import com.android.messaging.ui.contact.AddContactActivity
 import com.android.messaging.ui.conversation.LaunchConversationActivity
 import com.android.messaging.ui.license.LicenseActivity
-import com.android.messaging.ui.permissioncheck.PermissionCheckActivity
+import com.android.messaging.ui.onboarding.OnboardingActivity
 import com.android.messaging.ui.photoviewer.PhotoViewerActivity
 import java.io.File
 import javax.xml.parsers.DocumentBuilderFactory
@@ -24,7 +24,7 @@ internal class ActivityPermissionGateTest {
     )
 
     private val intentionallyUngated = setOf<Class<*>>(
-        PermissionCheckActivity::class.java,
+        OnboardingActivity::class.java,
         LaunchConversationActivity::class.java,
         PhotoViewerActivity::class.java,
         LicenseActivity::class.java,
