+++ src/com/android/messaging/domain/onboarding/model/PermissionRequest.kt
@@ -1,0 +1,14 @@
+package com.android.messaging.domain.onboarding.model
+
+import kotlinx.collections.immutable.ImmutableList
+
+internal sealed interface PermissionRequest {
+
+    data object AlreadyGranted : PermissionRequest
+
+    data object SmsRole : PermissionRequest
+
+    data class RuntimePermissions(
+        val permissions: ImmutableList<String>,
+    ) : PermissionRequest
+}
