+++ src/com/android/messaging/ui/BaseBugleActivity.java
@@ -32,7 +32,7 @@ public class BaseBugleActivity extends AppCompatActivity {
     @Override
     protected void onCreate(final Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        if (UiUtils.redirectToPermissionCheckIfNeeded(this)) {
+        if (UiUtils.redirectToOnboardingIfNeeded(this)) {
             return;
         }
     }
