+++ src/com/android/messaging/ui/conversation/LaunchConversationActivity.java
@@ -57,7 +57,7 @@ public class LaunchConversationActivity extends Activity implements
     @Override
     protected void onCreate(final Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        if (UiUtils.redirectToPermissionCheckIfNeeded(this)) {
+        if (UiUtils.redirectToOnboardingIfNeeded(this)) {
             return;
         }
