+++ src/com/android/messaging/ui/onboarding/screen/model/OnboardingAction.kt
@@ -1,0 +1,8 @@
+package com.android.messaging.ui.onboarding.screen.model
+
+internal sealed interface OnboardingAction {
+
+    data object NextClicked : OnboardingAction
+
+    data object SettingsClicked : OnboardingAction
+}
