+++ src/com/android/messaging/ui/onboarding/screen/model/SettingsGuidance.kt
@@ -1,0 +1,6 @@
+package com.android.messaging.ui.onboarding.screen.model
+
+internal enum class SettingsGuidance {
+    DefaultSmsApp,
+    Permissions,
+}
