--- src/com/android/messaging/ui/permissioncheck/screen/model/PermissionCheckAction.kt
@@ -1,8 +1,0 @@
-package com.android.messaging.ui.permissioncheck.screen.model
-
-internal sealed interface PermissionCheckAction {
-
-    data object NextClicked : PermissionCheckAction
-
-    data object SettingsClicked : PermissionCheckAction
-}
