+++ src/com/android/messaging/util/BuglePrefsKeys.java
@@ -62,6 +62,12 @@ public final class BuglePrefsKeys {
             = "conversation_notification_channels_migrated_v1";
 
     /**
+     * Whether the SMS insecurity warning has been acknowledged.
+     */
+    public static final String SMS_WARNING_ACKNOWLEDGED = "sms_warning_acknowledged";
+    public static final boolean SMS_WARNING_ACKNOWLEDGED_DEFAULT = false;
+
+    /**
      * The last selected chooser index in the media picker.
      */
     public static final String SELECTED_MEDIA_PICKER_CHOOSER_INDEX
